refactor: rename marks to markit
