+++ lua/markit/bookmark.lua
@@ -1,4 +1,4 @@
-local utils = require('marks.utils')
+local utils = require('markit.utils')
 local a = vim.api
 local Path = require('plenary.path')
